hexagonrpcd: iobuffer: use size_t for buffer sizes
The FastRPC peer can encode a 32-bit size for a variable-length
sequence. The members of the sequence can have a size more than 1,
potentially using more than 32 bits. Expand the buffer size to a size_t.

+++ hexagonrpcd/iobuffer.h
@@ -27,7 +27,7 @@
 #include <sys/types.h>
 
 struct fastrpc_io_buffer {
-	uint32_t s;
+	size_t s;
 	void *p;
 };
